requirements.txt
Django==1.7
Pillow==2.6.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-tastypie==0.12.1-dev
django-toolbelt==0.0.1
gunicorn==19.1.1
pilkit==1.1.12
psycopg2==2.5.4
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
sorl-thumbnail==11.12.1b
static3==0.5.1
